[Update] Studio Search - Added WithStudioSearch trait - Added filtering on is_nsfw attribute - Refactored studio/index search
